+++ CHANGELOG.md
@@ -11,7 +11,10 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
     * Create & Delete API integration with flow
 * Bugfix - User Invite - New User Role preselected fix
 * Bugfix - App User Management - Confirmation and User Role Changes
+* Shared Components
+   * More versatile image component
 * Feature - User Management - fetch roles (api refactoring)
+
 
 ## 0.7.0